+++ blakserv/ccode.c
@@ -173,7 +173,7 @@ blak_int C_Debug(int object_id,local_var_type *local_vars,
 				r = GetResourceByID(each_val.v.data);
 				if (r == NULL)
 				{
-					sprintf(buf+strlen(buf),"<unknown RESOURCE %lli>",each_val.v.data);
+					sprintf(buf+strlen(buf),"<unknown RESOURCE %lli>", (long long) each_val.v.data);
 				}
 				else
 				{
@@ -192,7 +192,7 @@ blak_int C_Debug(int object_id,local_var_type *local_vars,
 				c = GetClassByID(each_val.v.data);
 				if (c == NULL)
 				{
-					sprintf(buf+strlen(buf),"<unknown CLASS %lli>",each_val.v.data);
+					sprintf(buf+strlen(buf),"<unknown CLASS %lli>", (long long) each_val.v.data);
 				}
 				else
 				{
@@ -242,13 +242,13 @@ blak_int C_Debug(int object_id,local_var_type *local_vars,
 				o = GetObjectByID(each_val.v.data);
 				if (o == NULL)
 				{
-					sprintf(buf+strlen(buf),"<OBJECT %lli invalid>",each_val.v.data);
+					sprintf(buf+strlen(buf),"<OBJECT %lli invalid>",(long long) each_val.v.data);
 					break;
 				}
 				c = GetClassByID(o->class_id);
 				if (c == NULL)
 				{
-					sprintf(buf+strlen(buf),"<OBJECT %lli unknown class>",each_val.v.data);
+					sprintf(buf+strlen(buf),"<OBJECT %lli unknown class>",(long long) each_val.v.data);
 					break;
 				}
 				
@@ -258,10 +258,10 @@ blak_int C_Debug(int object_id,local_var_type *local_vars,
 					u = GetUserByObjectID(o->object_id);
 					if (u == NULL)
 					{
-						sprintf(buf+strlen(buf),"<OBJECT %lli broken user>",each_val.v.data);
+						sprintf(buf+strlen(buf),"<OBJECT %lli broken user>",(long long) each_val.v.data);
 						break;
 					}
-					sprintf(buf+strlen(buf),"ACCOUNT %i OBJECT %lli",u->account_id,each_val.v.data);
+					sprintf(buf+strlen(buf),"ACCOUNT %i OBJECT %lli",u->account_id,(long long) each_val.v.data);
 					break;
 				}
 			}
